-декомпозирует компоненты в main-page
