+++ src/components/cities/cities.tsx
@@ -1,32 +1,42 @@
+import { useState } from 'react';
+import { useAppSelector } from '../../hooks/use-app-selector';
 import { TOffer } from '../../types/offer';
 import { Map } from '../map/map';
 import { OfferList } from '../offer-list/offer-list';
 import { CitiesEmpty } from './cities-empty';
 
 type TCitiesProps = {
-  offers: TOffer[];
-  onOfferHover: (offerId: string) => void;
-  selectedPoint?: TOffer;
   locationCity: string;
 };
 
-const Cities = ({
-  offers,
-  selectedPoint,
-  locationCity,
-  onOfferHover,
-}: TCitiesProps) => {
-  const [currentOffer] = offers;
+const Cities = ({ locationCity }: TCitiesProps) => {
+  const offers = useAppSelector((store) => store.offers);
+  const [hoveredOffer, setHoveredOffer] = useState<TOffer | undefined>(
+    undefined
+  );
+  const currentLocationOffers = offers.filter(
+    (offer) => offer.city.name === locationCity
+  );
+  const [currentLocationOffer] = currentLocationOffers;
 
-  const citiesContent = offers.length ? (
+  const handleOfferHover = (offerId: string) => {
+    const currentOffer = offers.find((offer) => offer.id === offerId);
+
+    setHoveredOffer(currentOffer);
+  };
+
+  const citiesContent = currentLocationOffers.length ? (
     <div className="cities__places-container container">
-      <OfferList offers={offers} onOfferHover={onOfferHover} />
+      <OfferList
+        offers={currentLocationOffers}
+        onOfferHover={handleOfferHover}
+      />
       <div className="cities__right-section">
         <section className="cities__map map">
           <Map
-            city={currentOffer.city}
-            points={offers}
-            selectedPoint={selectedPoint}
+            city={currentLocationOffer.city}
+            points={currentLocationOffers}
+            selectedPoint={hoveredOffer}
           />
         </section>
       </div>
